gw-capture events
